services/leaderboards/pyproject.toml
[project]
name = "spacecases-leaderboards"
version = "0.1.0"
description = "SpaceCases leaderboard generatio scripts"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "python-dotenv==1.0.1",
    "aiohttp==3.11.11",
    "asyncpg==0.30.0",
    "asyncpg-stubs==0.30.0",
    "common"
]

[dependency-groups]
dev = [
    "ruff==0.8.4",
    "pytest==8.3.5",
    "pytest-custom-exit-code==0.3.0",
    "pyright==1.1.396",
]

[tool.uv.workspace]
members = ["../../common"]

[tool.uv.sources]
common = { workspace = true }
